requirements.txt
amqp==2.1.4
appdirs==1.4.3
Babel==2.3.4
billiard==3.5.0.2
celery==4.0.2
dj-database-url==0.4.1
Django==1.10.4
django-phonenumber-field==1.2.0
django-twilio==0.8.0
gunicorn==19.6.0
httplib2==0.10.3
kombu==4.0.2
packaging==16.8
phonenumberslite==8.3.3
psycopg2==2.6.2
pyparsing==2.2.0
PySocks==1.6.6
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
twilio==5.7.0
vine==1.1.3
whitenoise==3.2
